requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pandas>=1.3.0
pint>=0.18
scikit-learn>=0.24.0
plotly>=5.1.0
flask>=2.0.0
gunicorn>=20.1.0
numpy-stl>=2.16.0
trimesh>=3.9.0
vtk>=9.1.0
pytest>=6.2.0
black>=21.5b2
flake8>=3.9.0
mypy>=0.910
# TODO: Add NASA CEA or Cantera when implementing real gas effects
# rocketcea>=1.1.0  # For NASA CEA integration
# cantera>=3.0.0    # Alternative for real gas calculations 
gmsh>=4.9.0
cantera>=3.0.0
pygmsh>=7.1.0 